+++ lib/ts/recipe/webauthn/components/themes/mfa/loadingScreen.tsx
@@ -12,4 +12,3 @@ export const WebauthnMFALoadingScreen = withOverride("WebauthnMFALoadingScreen",
         </div>
     );
 });
-