+++ test/end-to-end/mfa.chooserscreen.test.js
@@ -24,6 +24,7 @@ import {
     screenshotOnFailure,
     getFactorChooserOptions,
     isMFASupported,
+    isWebauthnSupported,
     expectErrorThrown,
     setupBrowser,
     backendHook,